+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes to the export options:
 * Added a button in the main dialog to adjust export options previously only available when adding the `Export` procedure.
 * Renamed `Save in Folder` to `Folder` and `Save As` to `Name` for brevity.
 * By default, when a file about to be exported already exists with the same name, the new file will be renamed. This was inconsistent previously (e.g. when running non-interactively, the default was to skip the new files).
+* The default export folder is now the `Pictures` user folder instead of `Documents`.
 
 Changes to the `Export` procedure:
 * For Export Layers, the `Export` procedure now performs additional export instead of overriding the default export. The procedure name for Export Layers is now `Also export as...` to reflect this change. The name changes automatically as you modify the file extension (e.g. `Also export as PNG` if the file extension is `png`).